[helpers] move rule helpers into their own package under rules
