+++ rules/helpers/base.py
