+++ rules/sample_matchers.py
@@ -12,7 +12,7 @@ You can also supply multiple matchers for many common scenarios:
     @rule('root_logins', data_source=['osquery'],
           matchers=['prod', 'itx_corp'], sink=['csirt-pagerduty'])
 """
-from stream_alert.rule_helpers import in_set, last_hour
+from helpers.base import in_set, last_hour
 from stream_alert.rules_engine import StreamRules
 
 matcher = StreamRules.matcher
